+++ src/core/store/wallet/actions.js
@@ -26,6 +26,10 @@ const setTokens = function ({ commit }, params) {
   commit('SET_TOKENS', params);
 };
 
+const setCoinGeckoTokens = function ({ commit }, params) {
+  commit('SET_COIN_TOKENS', params);
+};
+
 const setAccountBalance = function ({ commit }, balance) {
   commit('SET_BALANCE', balance);
 };
@@ -121,5 +125,6 @@ export default {
   setWeb3Instance,
   setBlockNumber,
   setOwnedDomains,
-  setTokens
+  setTokens,
+  setCoinGeckoTokens
 };