+++ src/modules/access-wallet/hardware/handlers/hardwares/bitbox02/index.js
@@ -23,6 +23,10 @@ class BitBox02Wallet {
     this.supportedPaths = bip44Paths[WALLET_TYPES.BITBOX2];
     this.status = undefined;
     this.pairingConfirmed = false;
+    this.icon = {
+      type: 'mew-icon',
+      value: 'bitbox'
+    };
   }
   async connect() {
     const devicePath = await getDevicePath();
@@ -119,7 +123,8 @@ class BitBox02Wallet {
       errorHandler,
       txSigner,
       msgSigner,
-      displayAddress
+      displayAddress,
+      this.icon
     );
   }