+++ src/utils/networks/types/ETC.js
@@ -17,6 +17,8 @@ export default {
   isEthVMSupported: {
     supported: false,
     url: null,
+    blockExplorerTX: '',
+    blockExplorerAddr: '',
     websocket: null
   }
 };