+++ src/utils/networks/types/ETH.js
@@ -24,6 +24,8 @@ export default {
   isEthVMSupported: {
     supported: true,
     url: 'http://api.ethvm.com/',
+    blockExplorerTX: 'https://ethvm.com/tx/[[txHash]]',
+    blockExplorerAddr: 'https://ethvm.com/address/[[address]]',
     websocket: null
   }
 };