+++ src/utils/networks/types/GOERLI.js
@@ -24,6 +24,8 @@ export default {
   isEthVMSupported: {
     supported: false,
     url: null,
+    blockExplorerTX: '',
+    blockExplorerAddr: '',
     websocket: null
   }
 };